improved setup and play script.\setup will not start webiopi at boot anymore.\ndisplays back webiopi usage at end of install.\nplay script now start webiopi service before midori launch and stop it when closed.

+++ setup.sh
@@ -1,12 +1,16 @@
 #!/bin/sh
+# WebIOPi setup script
+
 SEARCH="python python3"
 FOUND=""
 INSTALLED=""
 
 echo "Installing WebIOPi...\n"
 
+# Install Python library
 cd python
 
+# Look up for installed python
 for python in $SEARCH; do
 	program="/usr/bin/$python"
 	if [ -x $program ]; then
@@ -15,15 +19,18 @@ for python in $SEARCH; do
 		include=`$python -c "import distutils.sysconfig; print(distutils.sysconfig.get_python_inc())"`
 		echo "Found $version... "
 
+		# Install required dev header if not present
 		if [ ! -f "$include/Python.h" ]; then
 			echo "Trying to install $python-dev using apt-get"
 			apt-get install -y "$python-dev"
 		fi
 
+		# Try to compile and install for the current python
 		if [ -f "$include/Python.h" ]; then
 			echo "Trying to install WebIOPi for $version"
 			$python setup.py install
 			if [ "$?" -ne "0" ]; then
+				# Sub setup error, continue with next python
 				echo "Build for $version failed\n"
 				continue
 			fi
@@ -35,8 +42,10 @@ for python in $SEARCH; do
 	fi
 done
 
+# Go back to the root folder
 cd ..
 
+# Ensure WebIOPi is installed to continue
 if [ -z "$INSTALLED" ]; then
 	if [ -z "$FOUND" ]; then
 		echo "ERROR: WebIOPi cannot be installed - neither python or python3 found"
@@ -47,27 +56,45 @@ if [ -z "$INSTALLED" ]; then
 	fi
 fi
 
+# Select greater python version
 for python in $INSTALLED; do
 	echo $python > /dev/null
 done 
 
+# Copy mandatory and optional resource files
+
+# Mandatory update
 echo "Copying resources..."
 mkdir /usr/share/webiopi 2>/dev/null 1>/dev/null
 cp -rf htdocs /usr/share/webiopi
+cp -rf python/webiopi-passwd.py /usr/bin/webiopi-passwd
+sed -i "s/python/$python/g" /usr/bin/webiopi-passwd
+
+# Add passwd file if it does not exist
 if [ ! -f "/etc/webiopi/passwd" ]; then
 	mkdir /etc/webiopi 2>/dev/null 1>/dev/null
 	cp python/passwd /etc/webiopi/passwd
 fi
-cp -rf python/webiopi-passwd.py /usr/bin/webiopi-passwd
-cp -rf python/webiopi.py.init /etc/init.d/webiopi
 echo
-echo "Setting up startup script..."
-chmod 0755 /etc/init.d/webiopi
-sed -i "s/python/$python/g" /etc/init.d/webiopi
-sed -i "s/python/$python/g" /usr/bin/webiopi-passwd
-update-rc.d webiopi defaults
+
+# Add service/daemon script if it does not exist
+if [ ! -f "/etc/init.d/webiopi" ]; then
+	echo "Setting up startup script..."
+	cp -rf python/webiopi.py.init /etc/init.d/webiopi
+	chmod 0755 /etc/init.d/webiopi
+	sed -i "s/python/$python/g" /etc/init.d/webiopi
+	echo
+fi
+
+# Display WebIOPi usages
+echo "WebIOPi successfully installed"
+for python in $INSTALLED; do
+        version=`$python -V 2>&1`
+        echo "* To start WebIOPi with $python\t: sudo $python -m webiopi"
+done
 echo
-echo "Starting WebIOPi..."
-/etc/init.d/webiopi restart
+echo "* To start WebIOPi at boot\t: sudo update-rc.d webiopi defaults"
+echo "* To start WebIOPi service\t: sudo /etc/init.d/webiopi start"
 echo
-echo "WebIOPi successfully installed, open a browser to http://localhost:8000/"
+echo "* Look in `pwd`/examples for Python library usage examples"
+echo
